Re-structure project with supplier state actions
> Uses 're-entrant' locking mechanism to prevent blocking in same thread

> Shift 'controller' and 'generator' in the 'supplier' module

+++ supplier/supplier.py
@@ -22,60 +22,105 @@ class Supplier:
 		self.slave = self.tables.difference(self.master)
 
 		self.count = 0
-		self.lock = threading.Lock()
+		self.lock = threading.RLock()
 
 		self.sample_size = sample_size
 		self.store_size = store_size
 		self.caution_threshold = caution_threshold
+
+		self.caution_aware = False
+		self.slave_ready = False
+
+		self.connect()
 
 	# connect the instance to the database
 	def connect(self):
 		try:
 			self.connection = r.connect(host='localhost', port=28015)
 
+			# create the 'id_archive' database if not present
+			db_list = r.db_list().run(self.connection)
+			if (DATABASE not in db_list):
+				r.db_create(DATABASE).run(self.connection)
+
 			# use the 'id_archive' database with this connection
-			r.db_create(DATABASE).run(self.connection)
 			self.connection.use(DATABASE)
 
 			# store reference to current 'table' and 'database'
 			self.db = r.db(DATABASE)
 			self.table = self.db.table(next(iter(self.master)))
+
+			# create 'storage' tables
+			self.create_tables()
 		except Exception, e:
 			raise e
 
 	# create all the required tables
 	def create_tables(self):
 		try:
+			# existing tables for the database
+			table_list = self.db.table_list().run(self.connection)
+
 			for table in self.tables:
-				r.table_create(table).run(self.connection)
+				if (table not in table_list):
+					r.table_create(table).run(self.connection)
 		except Exception, e:
 			raise e
 
 	# query a set of 'id' documents and remove them
 	def query_id(self):
+		# using GIL approach
+		self.lock.acquire()
+
+		# 'master' in 'danger' zone but 'slave' storage not ready
+		if (not self.slave_ready and self.master_state() == STATES['danger']):
+			self.lock.release()
+			return []
+
 		try:
-			# using GIL approach
-			with (self.lock):
-				# collect documents of given sample size
-				response = list(self.table.order_by('id').limit(self.sample_size).run(self.connection))
+			# collect documents of given sample size
+			response = list(self.table.order_by('id').limit(self.sample_size).run(self.connection))
 
-				# remove them from the database, makes use of the RethinkDB UUID keys
-				self.table.order_by('id').limit(self.sample_size).delete().run(self.connection)
+			# remove them from the database, makes use of the RethinkDB UUID keys
+			self.table.order_by('id').limit(self.sample_size).delete().run(self.connection)
 
-				# update the global 'counter' resource
-				self.count += len(response)
+			# update the global 'counter' resource
+			self.count += len(response)
 
-				return response
+			# regularly check for state change
+			self.state_check()
+
 		except Exception, e:
 			raise e
+		finally:
+			self.lock.release()
+
+		return response
 
 	# return the state of 'master' storage table
 	def master_state(self):
+		self.lock.acquire()
 		count = self.table.count().run(self.connection)
 
 		if (count < self.caution_threshold):
-			return STATES['safe']
+			state = STATES['safe']
 		elif (self.caution_threshold <= count < self.store_size):
-			return STATES['caution']
+			state = STATES['caution']
 		else:
-			return STATES['danger']
+			state = STATES['danger']
+
+		self.lock.release()
+		return state
+
+	# take further actions according to the current state
+	def state_check(self):
+		self.lock.acquire()
+		state = self.master_state()
+
+		if (state == STATES['caution']):
+			if (not self.caution_aware):
+				# ask 'controller' to start filling process [gevent]
+				pass
+		elif (state == STATES['danger']):
+			# shuffle the storage
+			pass
